api-tests/fastapi/requirements.txt
fastapi>=0.115.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
testcontainers>=4.8.0
psycopg2-binary>=2.9.9
